llnode: added commands to look at the work queue
Added two new commands (getactivehandles and getactiverequests)
which returns all pending handles and requests (same return
given by process._getActiveHandles() and process._getActiveRequests()).
This is a proof-of-concept implementation that currently only works
using the node debug build. Probably some modifications on node are
necessary to make it work on the official build.