src: refactor of workqueue related code
All code related to the workqueue commands were refactored by assigning
constants in a different file and by creating helper classes to access
some node internal properties.

Besides that, a new fallback method to determine the current environment
was added. This method uses a heuristic approach to try to find the
root context of the process. With this, those new commands will work
most of the time even without the new symbols added to V8.