build: split GitHub Actions
Split GitHub Actions to avoid `if` statements and to allow
cancelling/rerunning specific platforms. This simplifies run for macOS
as well as nightly runs. Also remove linter from daily runs, since lint
errors should be catched on PRs and commits.

Nightly runs now will fail on error, but will only run daily (not on PRs
or push). This way we still get a signal if llnode is working for  on
Node.js master or latest V8, but we don't pollute PRs and commits with
it.

These changes also make it easier to use docker images on Linux builds
in the future, to avoid installing lldb every time.